Added scaling and optimized file size more
Realized that when scanning in as columns, the width is always equal to 1, so I just leave that as a constant. Strangely seems to make some files larger, probably because it may not be as efficient for LZMA. Next step will be merging GLP and GEL formats, and auto-switching between them. Should help simplify things.